clip_omega argument to isvi, with stan submodule update

+++ src/cmdstan/arguments/arg_lambda.hpp
@@ -19,7 +19,7 @@ class arg_lambda : public real_argument {
     _value = _default_value;
   }
 
-  bool is_valid(int value) { return value >= 1.0; }
+  bool is_valid(double value) { return value >= 1.0; }
 };
 
 }  // namespace cmdstan
